tools/aft_aws_acess/pyproject.toml
[project]
name = "aft-aws-access-tools"
version = "0.1.0"
description = "Tools for comparing AWS AD groups against SSO configuration"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "ruff>=0.0.285",
    "mypy>=1.5.0",
    "types-pyyaml>=6.0.12",
]

[project.scripts]
compare-ad-groups = "compare_ad_groups:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = [
    "compare_ad_groups.py",
    "example_usage.py",
]

[tool.hatch.build.targets.sdist]
only-include = [
    "compare_ad_groups.py",
    "example_usage.py",
    "pyproject.toml",
    "README.md",
]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=compare_ad_groups",
    "--cov-report=term-missing",
]
